Update ann_cartpole.py
Added alternative functions for new_generation() and crossover_children(). Please remove the f1_ or f2_ for one of these to pick which one to use. The f1_ functions are the original ones in the master. 
The f2_new_generation uses something called elitism, which means it lets through the best performing ANNs from the previous generation unchanged, instead of a pair of new children. This seems to give better results.
The f2_crossover_children is just a different way of splitting the parent "genes" to give to the children ANNs. In the original method, the split is on the individual weights going from all the nodes to the next nodes. In the f2_crossover_children, the split is on all the weights going from one node, to the other nodes. It is a bit hard to explain in words, but is easy to illustrate. Anyway, this does not seem to make much difference, except it might be slightly worse and less similar to the typical description for a genetic algorithm, so it seems the original f1_ version works better for crossover_children.