New methods and updated partial_fit for every step
In addition to the extra functions described in the last update, I have updated the partial_fit in the predict_step function. It is now at the start and uses the data from the last step and the action taken in the last step. This ensures the step actually gives a reward before using partial_fit. It is still commented / # and not active in my script, so you have to activate it by removing the comment / #. In my runs, it makes the algorithm worse, as we discussed earlier. It still seems to be better to run just using the genetic algorithm, but with elitism (letting through the best ANNs).